Adds support for dept. filter in booking form's practitioner selector (#13845)
